backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.17.0",
    "drf-spectacular>=0.29.0",
    "groq>=1.1.2",
    "kombu>=5.6.2",
    "mysqlclient>=2.2.8",
    "pypdf>=6.9.2",
    "redis>=7.3.0",
    "requests>=2.33.0",
]
